fix: refactor and hook useClear
