+++ src/app/auth/google/login-success/page.tsx
@@ -1,5 +1,5 @@
 'use client'
-import { OauthLoginSuccess } from "@/features/auth/ui"
+import { OauthLoginSuccess } from '@/features/auth/ui'
 
 export default function OauthLoginSuccessPage() {
   return <OauthLoginSuccess />
