+++ src/app/auth/registration/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 import { SignupForm } from '@/features/auth/ui/signupForm'
-import SentEmailModal from '@/features/auth/ui/sentEmailModal/SentEmailModal'
+import { SentEmailModal } from '@/features/auth/ui'
 import { useState } from 'react'
 
 const Registration = () => {
