--- src/features/auth/api/authSlice.ts
@@ -1,13 +1,0 @@
-import { createSlice } from '@reduxjs/toolkit'
-
-const initialState= {
-  userId: null,
-};
-
-export const authSlice = createSlice({
-  name: 'auth',
-  initialState,
-  reducers:(create)=>({
-
-  })
-})
