+++ src/features/auth/ui/authGuard/authGuard.module.scss
@@ -1,0 +1,38 @@
+.loadingContainer {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  height: 100vh;
+  gap: 16px;
+  background-color: transparent;
+  padding: 32px;
+  animation: fadeIn 0.3s ease-in-out;
+}
+
+.loadingText {
+  font-size: 1rem;
+  color: var(--color-accent);
+  animation: pulseText 1.2s ease-in-out infinite;
+}
+
+@keyframes fadeIn {
+  from {
+    opacity: 0;
+  }
+  to {
+    opacity: 1;
+  }
+}
+
+@keyframes pulseText {
+  0% {
+    opacity: 0.7;
+  }
+  50% {
+    opacity: 1;
+  }
+  100% {
+    opacity: 0.7;
+  }
+}
