+++ src/features/auth/ui/authGuard/index.ts
@@ -1,0 +1,1 @@
+export * from './AuthGuard'
