+++ src/features/auth/ui/index.ts
@@ -2,3 +2,6 @@ export { OauthLoginSuccess } from './oauthLoginSuccess'
 export { EmailVerification } from './emailVerification'
 export { Login } from './login'
 export { SignupForm } from './signupForm'
+export { AuthGuard } from './authGuard'
+export { SentEmailModal } from './sentEmailModal'
+export { LogoutModal } from './logout'
