+++ src/features/auth/ui/sentEmailModal/index.ts
@@ -1,0 +1,1 @@
+export * from './SentEmailModal'
